separate_functions
