+++ lizrd/scripts/grid.py
@@ -6,28 +6,36 @@ Remember to set RUNNER and PARAMS in the script or add an argument parser.
 
 import argparse
 import datetime
-import pprint
 import subprocess
 import yaml
 from time import sleep
-
-from lizrd.scripts.grid_utils import (
-    create_grid,
-    get_train_main_function,
-    multiply_grid,
-    timestr_to_minutes,
-    get_machine_backend,
-    MachineBackend,
-    get_grid_entrypoint,
-    get_setup_args_with_defaults,
-    translate_to_argparse,
+from lizrd.grid.infrastructure import MachineBackend, get_machine_backend
+from lizrd.grid.prepare_configs import prepare_configs
+from lizrd.grid.setup_arguments import (
     make_singularity_env_arguments,
     make_singularity_mount_paths,
-    maybe_set_default_datasets_paths,
+)
+
+from lizrd.scripts.grid_utils import (
+    get_train_main_function,
+    timestr_to_minutes,
+    translate_to_argparse,
     check_for_argparse_correctness,
 )
+from lizrd.scripts.grid_utils import make_grid_from_configs
 from lizrd.support.code_copying import copy_code
-from lizrd.support.misc import load_with_inheritance
+import yaml
+
+
+def calculate_experiments_info(grid):
+    total_minutes = 0
+    total_n_experiments = 0
+    for setup_args, training_list in grid:
+        minutes_per_exp = timestr_to_minutes(setup_args["time"])
+        n_experiments = len(training_list)
+        total_n_experiments += n_experiments
+        total_minutes = n_experiments * minutes_per_exp
+    return total_minutes, total_n_experiments
 
 
 def create_subprocess_args(
@@ -39,53 +47,14 @@ def create_subprocess_args(
     skip_confirmation=False,
     skip_copy_code=False,
 ):
-    if config_path.endswith(".yaml"):
-        configs, all_config_paths = load_with_inheritance(config_path)
-    else:
-        raise ValueError("config path point to a .yaml")
-
-    for config in configs:
-        config["params"]["git_branch"] = git_branch
-        config["params"]["path_to_entry_config"] = config_path
-        config["params"]["all_config_paths"] = ",".join(all_config_paths)
-
-    interactive_options_per_config = [
-        config.get("interactive_debug", False) for config in configs
-    ]
-
-    assert (
-        len(set(interactive_options_per_config)) == 1
-    ), "`interactive_debug` must be the same for all configs"
-
-    interactive_debug_session = interactive_options_per_config[0]
-
-    # list of pairs: a dictionary of training_args and a dictionary of setup_args
-    grid = []
-    total_n_experiments = 0
-    total_minutes = 0
-
-    for i, config in enumerate(configs):
-        print(f"\nProcessing config {i}...")
-        pprint.pprint(config)
-        single_exp_training_args_grid = create_grid(config["params"])
-
-        setup_args = get_setup_args_with_defaults(config, CLUSTER_NAME)
-        single_exp_training_args_grid = multiply_grid(
-            single_exp_training_args_grid, setup_args["runs_multiplier"]
-        )
-        n_experiments = len(single_exp_training_args_grid)
-
-        grid += list(zip(single_exp_training_args_grid, [setup_args] * n_experiments))
-
-        total_n_experiments += n_experiments
-        minutes_per_exp = timestr_to_minutes(setup_args["time"])
-        total_minutes_from_this_grid = n_experiments * minutes_per_exp
-        total_minutes += total_minutes_from_this_grid
-
+    configs = prepare_configs(config_path, git_branch, CLUSTER_NAME)
+    grid = make_grid_from_configs(configs)
     check_for_argparse_correctness(grid)
+    interactive_debug_session = grid[0][0]["interactive_debug_session"]
 
     if CLUSTER_NAME != MachineBackend.LOCAL and not skip_confirmation:
         if not interactive_debug_session:
+            total_minutes, total_n_experiments = calculate_experiments_info(grid)
             user_input = input(
                 f"Will run {total_n_experiments} experiments, using up {total_minutes} minutes, i.e. around {round(total_minutes / 60)} hours\n"
                 f"Continue? [Y/n]"
@@ -99,109 +68,108 @@ def create_subprocess_args(
             exit(1)
 
     if CLUSTER_NAME != MachineBackend.LOCAL and (not skip_copy_code):
-        first_exp_training_args, _ = grid[0]
-        exp_name = first_exp_training_args["name"]
+        _, first_exp_trainings_args = grid[0]
+        exp_name = first_exp_trainings_args[0]["name"]
         newdir_name = (
             f"{exp_name}_{datetime.datetime.now().strftime('%Y-%m-%d-%H-%M-%S')}"
         )
         copy_code(newdir_name)
     else:
-        print("Running locally, skip copying code to a new directory.")
+        print("Skip copying code to a new directory.")
 
     slurm_command = "srun" if interactive_debug_session else "sbatch"
-    maybe_set_default_datasets_paths(grid, CLUSTER_NAME)
     experiments = []
-    for i, (training_args, setup_args) in enumerate(grid):
-        full_config_path = f"full_config{i}.yaml"
-        with open(full_config_path, "w") as f:
-            yaml.dump({**training_args, **setup_args}, f)
-        training_args["all_config_paths"] += f",{full_config_path}"
+    for setup_args, trainings_args in grid:
+        for i, training_args in enumerate(trainings_args):
+            full_config_path = f"full_config{i}.yaml"
+            with open(full_config_path, "w") as f:
+                yaml.dump({**training_args, **setup_args}, f)
+            training_args["all_config_paths"] += f",{full_config_path}"
 
-        job_name = training_args["name"]
-        training_args["n_gpus"] = setup_args["n_gpus"]
+            singularity_env_arguments = make_singularity_env_arguments(
+                hf_datasets_cache_path=setup_args["hf_datasets_cache"],
+                neptune_key=neptune_key,
+                wandb_key=wandb_key,
+            )
 
-        singularity_env_arguments = make_singularity_env_arguments(
-            hf_datasets_cache_path=setup_args["hf_datasets_cache"],
-            neptune_key=neptune_key,
-            wandb_key=wandb_key,
-        )
+            singularity_mount_paths = make_singularity_mount_paths(
+                setup_args, training_args
+            )
 
-        singularity_mount_paths = make_singularity_mount_paths(
-            setup_args, training_args
-        )
+            runner_params = translate_to_argparse(training_args)
+            if CLUSTER_NAME == MachineBackend.ENTROPY:
+                subprocess_args = [
+                    slurm_command,
+                    "--partition=a100",
+                    f"--gres=gpu:a100:{setup_args['n_gpus']}",
+                    f"--cpus-per-gpu={setup_args['cpus_per_gpu']}",
+                    f"--mem={1000 // setup_args['n_gpus']}G",
+                    f"--job-name={training_args['name']}",
+                    f"--time={setup_args['time']}",
+                    f"{setup_args['grid_entrypoint']}",
+                    "singularity",
+                    "run",
+                    *singularity_env_arguments,
+                    singularity_mount_paths,
+                    "--nv",
+                    setup_args["singularity_image"],
+                    "python3",
+                    "-m",
+                    setup_args["runner"],
+                    *runner_params,
+                ]
+            elif CLUSTER_NAME == MachineBackend.ATHENA:
+                subprocess_args = [
+                    slurm_command,
+                    f"--gres=gpu:{setup_args['n_gpus']}",
+                    "--partition=plgrid-gpu-a100",
+                    f"--cpus-per-gpu={setup_args['cpus_per_gpu']}",
+                    "--account=plgplggllmeffi-gpu-a100",
+                    f"--job-name={training_args['name']}",
+                    f"--time={setup_args['time']}",
+                    f"{setup_args['grid_entrypoint']}",
+                    "singularity",
+                    "run",
+                    "--bind=/net:/net",
+                    *singularity_env_arguments,
+                    singularity_mount_paths,
+                    "--nv",
+                    setup_args["singularity_image"],
+                    "python3",
+                    "-m",
+                    setup_args["runner"],
+                    *runner_params,
+                ]
+            elif CLUSTER_NAME == MachineBackend.IDEAS:
+                subprocess_args = [
+                    slurm_command,
+                    f"--gres=gpu:ampere:{setup_args['n_gpus']}",
+                    f"--cpus-per-gpu={setup_args['cpus_per_gpu']}",
+                    f"--job-name={training_args['name']}",
+                    f"--time={setup_args['time']}",
+                    "--mem=32G",
+                    setup_args["nodelist"],
+                    f"{setup_args['grid_entrypoint']}",
+                    "singularity",
+                    "run",
+                    *singularity_env_arguments,
+                    singularity_mount_paths,
+                    "--nv",
+                    setup_args["singularity_image"],
+                    "python3",
+                    "-m",
+                    setup_args["runner"],
+                    *runner_params,
+                ]
+            elif CLUSTER_NAME == MachineBackend.LOCAL:
+                runner_main_function = get_train_main_function(setup_args["runner"])
+                return [
+                    (runner_main_function, runner_params)
+                ], interactive_debug_session
+            else:
+                raise ValueError(f"Unknown cluster name: {CLUSTER_NAME}")
 
-        runner_params = translate_to_argparse(training_args)
-        if CLUSTER_NAME == MachineBackend.ENTROPY:
-            subprocess_args = [
-                slurm_command,
-                "--partition=a100",
-                f"--gres=gpu:a100:{setup_args['n_gpus']}",
-                f"--cpus-per-gpu={setup_args['cpus_per_gpu']}",
-                f"--mem={1000 // setup_args['n_gpus']}G",
-                f"--job-name={job_name}",
-                f"--time={setup_args['time']}",
-                get_grid_entrypoint(CLUSTER_NAME),
-                "singularity",
-                "run",
-                *singularity_env_arguments,
-                singularity_mount_paths,
-                "--nv",
-                setup_args["singularity_image"],
-                "python3",
-                "-m",
-                setup_args["runner"],
-                *runner_params,
-            ]
-        elif CLUSTER_NAME == MachineBackend.ATHENA:
-            subprocess_args = [
-                slurm_command,
-                f"--gres=gpu:{setup_args['n_gpus']}",
-                "--partition=plgrid-gpu-a100",
-                f"--cpus-per-gpu={setup_args['cpus_per_gpu']}",
-                "--account=plgplggllmeffi-gpu-a100",
-                f"--job-name={job_name}",
-                f"--time={setup_args['time']}",
-                get_grid_entrypoint(CLUSTER_NAME),
-                "singularity",
-                "run",
-                "--bind=/net:/net",
-                *singularity_env_arguments,
-                singularity_mount_paths,
-                "--nv",
-                setup_args["singularity_image"],
-                "python3",
-                "-m",
-                setup_args["runner"],
-                *runner_params,
-            ]
-        elif CLUSTER_NAME == MachineBackend.IDEAS:
-            subprocess_args = [
-                slurm_command,
-                f"--gres=gpu:ampere:{setup_args['n_gpus']}",
-                f"--cpus-per-gpu={setup_args['cpus_per_gpu']}",
-                f"--job-name={job_name}",
-                f"--time={setup_args['time']}",
-                "--mem=32G",
-                setup_args["nodelist"],
-                get_grid_entrypoint(CLUSTER_NAME),
-                "singularity",
-                "run",
-                *singularity_env_arguments,
-                singularity_mount_paths,
-                "--nv",
-                setup_args["singularity_image"],
-                "python3",
-                "-m",
-                setup_args["runner"],
-                *runner_params,
-            ]
-        elif CLUSTER_NAME == MachineBackend.LOCAL:
-            runner_main_function = get_train_main_function(setup_args["runner"])
-            return [(runner_main_function, runner_params)], interactive_debug_session
-        else:
-            raise ValueError(f"Unknown cluster name: {CLUSTER_NAME}")
-
-        experiments.append((subprocess_args, job_name))
+            experiments.append((subprocess_args, training_args["name"]))
     return experiments, interactive_debug_session
